shopping-cart
I have implemented the "Add to cart" functionality using redux-toolkit and local storage. As I have created a context and wrapped the app around the provider of that context so the product is added to cart without reloading of the page. And local storage maintains the cart items state in the storage even when we reload the app. Upon adding same items just the quantity of the already present item is updated so we don't have repeated items. It could be done by just updating the local storage when adding or updating any changes in cart but for that we had to reload the page that will re render each component and the performance of the app will be compromised that's why I used redux-toolkit for its setup.